+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 - Fix Drop component no longer flips, if it does not fit into scroll area
 - Fix Drop component alignment within Dropbar
 - Fix ensure write updates aren't called if component is already disconnected
+- Fix Offcanvas component no longer overflows document if isn't wide enough
 - Remove margin from the last-child within dropdowns in Dropdown and Navbar components
 
 ## 3.14.3 (May 27, 2022)